+++ apps/native-component-list/src/screens/UI/BottomSheetScreen.android.tsx
@@ -1,12 +1,25 @@
+import { Button, BottomSheet } from '@expo/ui/jetpack-compose';
 import * as React from 'react';
 import { ScrollView, Text, View } from 'react-native';
 
 export default function BottomSheetScreen() {
+  const [isOpened, setIsOpened] = React.useState<boolean>(false);
+
   return (
-    <ScrollView style={{ flex: 1 }}>
-      <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-        <Text style={{ textAlign: 'center' }}>Not implemented yet on Android</Text>
-      </View>
+    <ScrollView
+      contentContainerStyle={{
+        flex: 1,
+        justifyContent: 'flex-start',
+        alignItems: 'flex-start',
+        padding: 8,
+      }}>
+      <Button onPress={() => setIsOpened((h) => !h)}>Toggle</Button>
+      <Text>isOpened: {isOpened ? 'yes' : 'no'}</Text>
+      <BottomSheet isOpened={isOpened} onIsOpenedChange={(e) => setIsOpened(e)}>
+        <View style={{ padding: 20 }}>
+          <Text>Hello world</Text>
+        </View>
+      </BottomSheet>
     </ScrollView>
   );
 }